Dagster/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alabaster==1.0.0
alembic==1.17.2
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
anyio==4.11.0
app-model==0.4.0
appdirs==1.4.4
asttokens==3.0.1
attrs==25.4.0
babel==2.17.0
backoff==2.2.1
bermuda==0.1.6
build==1.3.0
cachey==0.2.1
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
coloredlogs==14.0
comm==0.2.3
connected-components-3d==3.26.1
dagit==1.12.3
dagster==1.12.3
dagster-graphql==1.12.3
dagster-pandas==0.28.3
dagster-pipes==1.12.3
dagster-shell==0.11.14
dagster-webserver==1.12.3
dagster_shared==1.12.3
dask==2025.11.0
dataclasses==0.6
debugpy==1.8.17
decorator==5.2.1
docstring_parser==0.17.0
docutils==0.21.2
donfig==0.8.1.post1
et_xmlfile==2.0.0
executing==2.2.1
filelock==3.20.0
flexcache==0.3
flexparser==0.4
freetype-py==2.5.1
frozenlist==1.8.0
fsspec==2025.10.0
google-crc32c==1.7.1
gql==3.5.3
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-health-checking==1.76.0
h11==0.16.0
HeapDict==1.0.1
hsluv==5.0.4
httptools==0.7.1
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
imagesize==1.4.1
importlib_metadata==8.7.0
importlib_resources==6.5.2
in-n-out==0.2.1
ipykernel==6.31.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
itk==5.4.5
itk-core==5.4.5
itk-filtering==5.4.5
itk-io==5.4.5
itk-numerics==5.4.5
itk-registration==5.4.5
itk-segmentation==5.4.5
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
lazy_loader==0.4
llvmlite==0.45.1
locket==1.0.0
loguru==0.7.3
magicgui==0.10.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
multidict==6.7.0
napari==0.6.6
napari-console==0.1.4
napari-plugin-engine==0.2.0
napari-plugin-manager==0.1.8
napari-svg==0.2.1
nest-asyncio==1.6.0
networkx==3.6
nibabel==5.3.2
NodeGraphQt==0.6.43
npe2==0.7.9
numba==0.62.1
numcodecs==0.16.5
numpy==2.2.6
numpydoc==1.9.0
opencv-python==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
parso==0.8.5
partd==1.4.2
pathlib_abc==0.5.2
pillow==12.0.0
Pint==0.25.2
platformdirs==4.5.0
pooch==1.8.2
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.1
psutil==7.1.3
psygnal==0.15.0
pure_eval==0.2.3
pyconify==0.2.1
pydantic==2.12.4
pydantic-compat==0.1.2
pydantic_core==2.41.5
pydicom==3.0.1
Pygments==2.19.2
pynrrd==1.1.3
PyOpenGL==3.1.10
pyproject_hooks==1.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.1
pyreadline3==3.5.4
pysera==2.1.5
PySide6==6.10.1
PySide6_Addons==6.10.1
PySide6_Essentials==6.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
Qt.py==1.4.8
qtconsole==5.7.0
QtPy==2.4.3
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
roman-numerals-py==3.1.0
rpds-py==0.29.0
rt-utils==1.2.7
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.3
shellingham==1.5.4
shiboken6==6.10.1
simpleitk==2.5.3
six==1.17.0
sniffio==1.3.1
snowballstemmer==3.0.1
Sphinx==8.2.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.44
stack-data==0.6.3
starlette==0.50.0
structlog==25.5.0
superqt==0.7.6
tabulate==0.9.0
threadpoolctl==3.6.0
tifffile==2025.10.16
tomli==2.3.0
tomli_w==1.2.0
tomlkit==0.13.3
toolz==1.1.0
toposort==1.10
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
triangle==20250106
typer==0.20.0
types-PySide2==5.15.2.1.8
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
universal_pathlib==0.3.6
urllib3==2.5.0
uvicorn==0.38.0
vispy==0.15.2
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
win32_setctime==1.2.0
wrapt==2.0.1
yarl==1.22.0
zarr==3.1.5
zipp==3.23.0
